Switch over name to simpler format and add a little more docs
