+++ lib/cucumber/events/test_run_finished.rb
@@ -3,7 +3,7 @@
 module Cucumber
   module Events
     # Event fired after all test cases have finished executing
-    class TestRunFinished < BaseEventNew
+    class TestRunFinished < Base
       attr_reader :success
 
       def self.event_id
