+++ lib/cucumber/events/test_run_started.rb
@@ -1,12 +1,12 @@
 # frozen_string_literal: true
 
-require_relative 'base_event_new'
+require_relative 'base'
 
 module Cucumber
   module Events
     # Event fired once all test cases have been filtered before
     # the first one is executed.
-    class TestRunStarted < BaseEventNew
+    class TestRunStarted < Base
       # @return [Array<Cucumber::Core::Test::Case>] the test cases to be executed
       attr_reader :test_cases
