+++ lib/cucumber/events/test_step_started.rb
@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 
-require_relative 'base_event_new'
+require_relative 'base'
 
 module Cucumber
   module Events
     # Signals that a {Cucumber::Core::Test::Step} is about to be executed
-    class TestStepStarted < BaseEventNew
+    class TestStepStarted < Base
       # @return [Cucumber::Core::Test::Step] the test step to be executed
       attr_reader :test_step
