+++ lib/cucumber/glue/registry_and_more.rb
@@ -13,7 +13,7 @@ require 'cucumber/glue/world_factory'
 require 'cucumber/gherkin/i18n'
 require 'multi_test'
 require 'cucumber/step_match'
-require 'cucumber/events/base_event_new'
+require 'cucumber/events/base'
 require 'cucumber/events/step_definition_registered'
 
 module Cucumber
